Fix invalid test-input file
This file generated a libical error complaining

    Can't parse as RECUR value in RRULE property. Removing entire property: FREQ=YEARLY\;UNTIL=123456T123456\;BYSETPOS=-1\,2\,-3\,4\,-5\,6\,-7\,8\,9\,10\,11\,12\,13\,14\,15\,16\,17\,18\,19\,20\,21\,22\,23\,4\,25\,26\,27\,28\,29\,30\,31\,32\,33\,34\,35\,36\,37\,38\,39\,40\,41\,42\,43\,44\,45\,46\,47\,48\,49\,50\,51\,52\,53\,54\;BYSECOND=1\,2\,3\,4\,5\,6\,7\,8\,9\,10\,11\,12\,13\,14\,15\,16\,17\,18\,19\,20\,21\,22\,23\,4\,25\,26

The error makes three of the tests for this file fail.

The issue appears to be that the UNTIL value is invalid; adding two more
digits (allowing it to be parsed as a YYYYMMDD string) fixes the error.
As far as I can tell this error is irrelevant for the affected tests.